Added tabs & more!
On this update, tabs were added they are located on top of the navigation bar this took me some time at least to learn how to do it correctly, also blocked the navigation bar so you can't drag it arround which was present on the previous version, also added a new style that made buttons bigger and a better UI in general. Also added a way to close the tabs created, and you can drag them to re-order them. In next updates you will be able to pin them.

Also deleted some un-used imports and added a resize fo the window that doesn't work very well but makes a good job on keeping the window usable at the start. Without it, the window is very small.