Another bug fix
When you were left with only 1 tab and deleted that one you will have an empty blank program, this modification checks for only 1 tab and in case that last is deleted it will create another. In the future, this will be an option so if you close the last tab you can choose to either close the program or create a new tab.